Migrated spatial bin hashes to uint64

+++ Source/PCGExtendedToolkit/Private/Topology/PCGExTopology.cpp
@@ -560,8 +560,8 @@ void FPCGExCellArtifactsDetails::Process(
 	}
 
 	if (bWriteCellHash) { InDataFacade->GetWritable<int64>(CellHashAttributeName, static_cast<int64>(InCell->GetCellHash()), true, PCGExData::EBufferInit::New); }
-	if (bWriteArea) { InDataFacade->GetWritable<double>(AreaAttributeName, static_cast<double>(InCell->Data.Area), true, PCGExData::EBufferInit::New); }
-	if (bWriteCompactness) { InDataFacade->GetWritable<double>(CompactnessAttributeName, static_cast<double>(InCell->Data.Compactness), true, PCGExData::EBufferInit::New); }
+	if (bWriteArea) { InDataFacade->GetWritable<double>(AreaAttributeName, InCell->Data.Area, true, PCGExData::EBufferInit::New); }
+	if (bWriteCompactness) { InDataFacade->GetWritable<double>(CompactnessAttributeName, InCell->Data.Compactness, true, PCGExData::EBufferInit::New); }
 	if (TerminalBuffer) { for (int i = 0; i < NumNodes; i++) { TerminalBuffer->SetValue(i, InCluster->GetNode(InCell->Nodes[i])->IsLeaf()); } }
 	if (RepeatBuffer) { for (int i = 0; i < NumNodes; i++) { RepeatBuffer->SetValue(i, NumRepeats[InCell->Nodes[i]] - 1); } }
